requirements.txt
pytest>=8.0.0
python-chess>=1.10.3
